feat：权限系统熔断设计 前置数据准备 #11789
